Added OuterRadii Rotation Curve + Removed Comments
Adding comments of the form "(comments here)" causes the line to fail to
parse. Common results: moving to "stow" or doing n-point scans instead
of track scans.  Removed all comments just to be safe.

Also, added script to measure redshifts at radii beyond that of the LSR.
Lab guide indicates that we can't extract information, but I want to get
data in case it is useful.